tools/llm/vllm/scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "vLLM scripts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "autoawq>=0.2.9",
  "transformers>=4",
  "torch>=2.9.1",
  "accelerate",
]
